rename data grid decorator
fixes casing issue
